Better reduction-based parallel heap construction.
