+++ include/minocore/hash/hash.h
@@ -77,7 +77,7 @@ public:
     }
 };
 
-template<typename FT=double, bool SO=blz::rowMajor>
+template<typename FT=double, bool SO=blaze::rowMajor>
 class JSDLSHasher {
     // See https://papers.nips.cc/paper/9195-locality-sensitive-hashing-for-f-divergences-mutual-information-loss-and-beyond.pdf
     // for the function.
@@ -85,8 +85,8 @@ class JSDLSHasher {
     //
     // This relies on a U/H-approximation of the JSD (capacitory discrimination in Topsoe, 2000)
     // by the Hellinger distance, and uses an LSH for the Hellinger as-is.
-    blz::DM<FT, SO> randproj_;
-    blz::DV<FT, SO> boffsets_;
+    blaze::DynamicMatrix<FT, SO> randproj_;
+    blaze::DynamicVector<FT, SO> boffsets_;
     LSHasherSettings settings_;
 public:
     using ElementType = FT;
@@ -96,33 +96,33 @@ public:
         if(seed == 0) seed = nd * nh + r;
         std::mt19937_64 mt(seed);
         std::normal_distribution<FT> gen;
-        randproj_ = blz::generate(nh, nd, [&](size_t x, size_t y){
+        randproj_ = blaze::generate(nh, nd, [&](size_t x, size_t y){
             std::mt19937_64 mt(seed + x + seed * y);
             return gen(mt);
         });
         randproj_ /= r;
-        boffsets_ = blz::generate(nh, [&](size_t){return FT(mt()) / mt.max();});
+        boffsets_ = blaze::generate(nh, [&](size_t){return FT(mt()) / mt.max();});
         assert(settings_.k_ * settings_.l_ == randproj_.rows()); // In case of overflow, I suppose
     }
     template<typename VT>
-    decltype(auto) hash(const blz::Vector<VT, SO> &input) const {
+    decltype(auto) hash(const blaze::Vector<VT, SO> &input) const {
         //std::fprintf(stderr, "Regular input size: %zu. my rows/col:%zu/%zu\n", (~input).size(), randproj_.rows(), randproj_.columns());
-        return blz::ceil(randproj_ * blz::sqrt(~input) + boffsets_);
+        return blaze::ceil(randproj_ * blaze::sqrt(~input) + boffsets_);
     }
     template<typename VT>
-    decltype(auto) hash(const blz::Vector<VT, !SO> &input) const {
+    decltype(auto) hash(const blaze::Vector<VT, !SO> &input) const {
         //std::fprintf(stderr, "Reversed input size: %zu. my rows/col:%zu/%zu\n", (~input).size(), randproj_.rows(), randproj_.columns());
-        return blz::ceil(randproj_ * trans(blz::sqrt(~input)) + boffsets_);
+        return blaze::ceil(randproj_ * trans(blaze::sqrt(~input)) + boffsets_);
     }
     template<typename VT>
-    decltype(auto) hash(const blz::Matrix<VT, SO> &input) const {
+    decltype(auto) hash(const blaze::Matrix<VT, SO> &input) const {
         //std::fprintf(stderr, "Regular input rows/col: %zu/%zu. my rows/col:%zu/%zu\n", (~input).rows(), (~input).columns(), randproj_.rows(), randproj_.columns());
-        return trans(blz::ceil(randproj_ * trans(blz::sqrt(~input)) + blz::expand(boffsets_, (~input).rows())));
+        return trans(blaze::ceil(randproj_ * trans(blaze::sqrt(~input)) + blaze::expand(boffsets_, (~input).rows())));
     }
     template<typename VT>
-    decltype(auto) hash(const blz::Matrix<VT, !SO> &input) const {
+    decltype(auto) hash(const blaze::Matrix<VT, !SO> &input) const {
         //std::fprintf(stderr, "Reversed SO input rows/col: %zu/%zu. my rows/col:%zu/%zu\n", (~input).rows(), (~input).columns(), randproj_.rows(), randproj_.columns());
-        return trans(blz::ceil(randproj_ * blz::sqrt(~input) + blz::expand(boffsets_, (~input).columns())));
+        return trans(blaze::ceil(randproj_ * blaze::sqrt(~input) + blaze::expand(boffsets_, (~input).columns())));
     }
     const auto &matrix() const {return randproj_;}
     auto dim() const {return randproj_.columns();}
@@ -131,7 +131,7 @@ public:
     auto l()   const {return settings_.l_;}
     const auto &settings() const {return settings_;}
 };
-template<typename FT=double, bool SO=blz::rowMajor>
+template<typename FT=double, bool SO=blaze::rowMajor>
 class HellingerLSHasher: public JSDLSHasher<FT, SO> {
 public:
     template<typename...Args>
@@ -145,8 +145,8 @@ class PStableLSHasher {
     // https://aaai.org/ocs/index.php/AAAI/AAAI17/paper/view/14692
     // for the derivation
     // Note that this is an LSH for the JS Metric, not the JSD.
-    blz::DM<FT, SO> randproj_;
-    blz::DV<FT, SO> boffsets_;
+    blaze::DynamicMatrix<FT, SO> randproj_;
+    blaze::DynamicVector<FT, SO> boffsets_;
     LSHasherSettings settings_;
     double w_;
 public:
@@ -161,34 +161,34 @@ public:
         auto nd = settings.dim_;
         if(seed == 0) seed = nd * nh  + w + 1. / w;
         std::mt19937_64 mt(seed);
-        randproj_ = blz::abs(blz::generate(nh, nd, [&](size_t, size_t){return gen(mt);}) * (1. / w));
+        randproj_ = blaze::abs(blaze::generate(nh, nd, [&](size_t, size_t){return gen(mt);}) * (1. / w));
         if constexpr(use_offsets)
-            boffsets_ = blz::generate(nh, [&](size_t){return FT(mt() / 2) / mt.max();}) - 0.5;
+            boffsets_ = blaze::generate(nh, [&](size_t){return FT(mt() / 2) / mt.max();}) - 0.5;
         assert(settings_.k_ * settings_.l_ == randproj_.rows()); // In case of overflow, I suppose
     }
     template<typename VT>
-    decltype(auto) hash(const blz::Vector<VT, SO> &input) const {
-        if constexpr(use_offsets) return blz::floor(randproj_ * (~input) + 1.) + boffsets_;
-        else                      return blz::floor(randproj_ * (~input));
+    decltype(auto) hash(const blaze::Vector<VT, SO> &input) const {
+        if constexpr(use_offsets) return blaze::floor(randproj_ * (~input) + 1.) + boffsets_;
+        else                      return blaze::floor(randproj_ * (~input));
     }
     template<typename VT>
-    decltype(auto) hash(const blz::Vector<VT, !SO> &input) const {
-        if constexpr(use_offsets) return blz::floor(randproj_ * trans(~input) + 1.) + boffsets_;
-        else                      return blz::floor(randproj_ * trans(~input));
+    decltype(auto) hash(const blaze::Vector<VT, !SO> &input) const {
+        if constexpr(use_offsets) return blaze::floor(randproj_ * trans(~input) + 1.) + boffsets_;
+        else                      return blaze::floor(randproj_ * trans(~input));
     }
     template<typename MT>
-    decltype(auto) hash(const blz::Matrix<MT, SO> &input) const {
+    decltype(auto) hash(const blaze::Matrix<MT, SO> &input) const {
         if constexpr(use_offsets)
-            return trans(blz::floor(randproj_ * trans(~input)) + blz::expand(boffsets_, (~input).rows()));
+            return trans(blaze::floor(randproj_ * trans(~input)) + blaze::expand(boffsets_, (~input).rows()));
         else
-            return trans(blz::floor(randproj_ * trans(~input)));
+            return trans(blaze::floor(randproj_ * trans(~input)));
     }
     template<typename MT>
-    decltype(auto) hash(const blz::Matrix<MT, !SO> &input) const {
+    decltype(auto) hash(const blaze::Matrix<MT, !SO> &input) const {
         if constexpr(use_offsets)
-            return trans(blz::floor(randproj_ * trans(~input)) + blz::expand(boffsets_, (~input).columns()));
+            return trans(blaze::floor(randproj_ * trans(~input)) + blaze::expand(boffsets_, (~input).columns()));
         else
-            return trans(blz::floor(randproj_ * trans(~input)));
+            return trans(blaze::floor(randproj_ * trans(~input)));
     }
     const auto &matrix() const {return randproj_;}
     auto dim() const {return settings_.dim_;}
@@ -198,7 +198,7 @@ public:
     const auto &settings() const {return settings_;}
 };
 
-template<typename FT=double, bool SO=blz::rowMajor, bool use_offsets=true>
+template<typename FT=double, bool SO=blaze::rowMajor, bool use_offsets=true>
 class L2LSHasher: public PStableLSHasher<std::normal_distribution, FT, SO, use_offsets> {
 public:
     using super = PStableLSHasher<std::normal_distribution, FT, SO, use_offsets>;
@@ -208,7 +208,7 @@ public:
     }
 };
 
-template<typename FT=double, bool SO=blz::rowMajor, bool use_offsets=true>
+template<typename FT=double, bool SO=blaze::rowMajor, bool use_offsets=true>
 class L1LSHasher: public PStableLSHasher<std::cauchy_distribution, FT, SO, use_offsets> {
 public:
     using super = PStableLSHasher<std::cauchy_distribution, FT, SO, use_offsets>;
@@ -217,7 +217,7 @@ public:
     {
     }
 };
-template<typename FT=double, bool SO=blz::rowMajor, bool use_offsets=true>
+template<typename FT=double, bool SO=blaze::rowMajor, bool use_offsets=true>
 class LpLSHasher: public PStableLSHasher<cms_distribution, FT, SO, use_offsets> {
 public:
     using super = PStableLSHasher<cms_distribution, FT, SO, use_offsets>;
@@ -227,7 +227,7 @@ public:
     }
 };
 
-template<typename FT=double, bool SO=blz::rowMajor, bool use_offsets=true>
+template<typename FT=double, bool SO=blaze::rowMajor, bool use_offsets=true>
 class ClippedL1LSHasher: public PStableLSHasher<clipped_cauchy_distribution, FT, SO, use_offsets> {
 public:
     using super = PStableLSHasher<clipped_cauchy_distribution, FT, SO, use_offsets>;
@@ -235,7 +235,7 @@ public:
     ClippedL1LSHasher(Args &&...args): super(std::forward<Args>(args)...) {}
 };
 
-template<typename FT=double, bool SO=blz::rowMajor, bool use_offsets=true>
+template<typename FT=double, bool SO=blaze::rowMajor, bool use_offsets=true>
 class TVDLSHasher: public L1LSHasher<FT, SO, use_offsets> {
 public:
     using super = L1LSHasher<FT, SO, use_offsets>;
@@ -244,14 +244,14 @@ public:
 };
 
 
-template<typename FT=double, bool SO=blz::rowMajor>
+template<typename FT=double, bool SO=blaze::rowMajor>
 class S2JSDLSHasher {
     // See S2JSD-LSH: A Locality-Sensitive Hashing Schema for Probability Distributions
     // https://aaai.org/ocs/index.php/AAAI/AAAI17/paper/view/14692
     // for the derivation
     // Note that this is an LSH for the JS Metric, not the JSD.
-    blz::DM<FT, SO> randproj_;
-    blz::DV<FT, SO> boffsets_;
+    blaze::DynamicMatrix<FT, SO> randproj_;
+    blaze::DynamicVector<FT, SO> boffsets_;
     LSHasherSettings settings_;
     double w_;
 public:
@@ -263,25 +263,25 @@ public:
         if(seed == 0) seed = nd * nh  + w + 1. / w;
         std::mt19937_64 mt(seed);
         std::normal_distribution<FT> gen;
-        randproj_ = blz::abs(blz::generate(nh, nd, [&](size_t, size_t){return gen(mt);}) * (4. / (w * w)));
-        boffsets_ = blz::generate(nh, [&](size_t){return FT(mt() / 2) / mt.max();}) - 0.5;
+        randproj_ = blaze::abs(blaze::generate(nh, nd, [&](size_t, size_t){return gen(mt);}) * (4. / (w * w)));
+        boffsets_ = blaze::generate(nh, [&](size_t){return FT(mt() / 2) / mt.max();}) - 0.5;
         assert(settings_.k_ * settings_.l_ == randproj_.rows()); // In case of overflow, I suppose
     }
     template<typename VT>
-    decltype(auto) hash(const blz::Vector<VT, SO> &input) const {
-        return blz::floor(blz::sqrt(randproj_ * (~input) + 1.) + boffsets_);
+    decltype(auto) hash(const blaze::Vector<VT, SO> &input) const {
+        return blaze::floor(blaze::sqrt(randproj_ * (~input) + 1.) + boffsets_);
     }
     template<typename VT>
-    decltype(auto) hash(const blz::Vector<VT, !SO> &input) const {
-        return blz::floor(blz::sqrt(randproj_ * trans(~input) + 1.) + boffsets_);
+    decltype(auto) hash(const blaze::Vector<VT, !SO> &input) const {
+        return blaze::floor(blaze::sqrt(randproj_ * trans(~input) + 1.) + boffsets_);
     }
     template<typename MT>
-    decltype(auto) hash(const blz::Matrix<MT, SO> &input) const {
-        return trans(blz::floor(blz::sqrt(randproj_ * trans(~input) + 1.) + blz::expand(boffsets_, (~input).rows())));
+    decltype(auto) hash(const blaze::Matrix<MT, SO> &input) const {
+        return trans(blaze::floor(blaze::sqrt(randproj_ * trans(~input) + 1.) + blaze::expand(boffsets_, (~input).rows())));
     }
     template<typename MT>
-    decltype(auto) hash(const blz::Matrix<MT, !SO> &input) const {
-        return trans(blz::floor(blz::sqrt(randproj_ * (trans(~input)) + 1.) + blz::expand(boffsets_, (~input).columns())));
+    decltype(auto) hash(const blaze::Matrix<MT, !SO> &input) const {
+        return trans(blaze::floor(blaze::sqrt(randproj_ * (trans(~input)) + 1.) + blaze::expand(boffsets_, (~input).columns())));
     }
     const auto &matrix() const {return randproj_;}
     auto dim() const {return settings_.dim_;}
@@ -370,8 +370,8 @@ public:
         return hasher_.hash(q);
     }
     template<typename VT, bool OSO>
-    void add(const blz::Vector<VT, OSO> &input, IT id) {
-        auto hv = blz::evaluate(hash(input));
+    void add(const blaze::Vector<VT, OSO> &input, IT id) {
+        auto hv = blaze::evaluate(hash(input));
         if(unlikely(nh_ != hv.size())) {
             std::fprintf(stderr, "[%s] nh_: %u. hv.size: %zu\n", __PRETTY_FUNCTION__, nh_, hv.size());
             std::exit(1);
@@ -383,8 +383,8 @@ public:
         }
     }
     template<typename MT, bool OSO>
-    void add(const blz::Matrix<MT, OSO> &input, IT idoffset=0) {
-        auto hv = blz::evaluate(hash(input));
+    void add(const blaze::Matrix<MT, OSO> &input, IT idoffset=0) {
+        auto hv = blaze::evaluate(hash(input));
         std::fprintf(stderr, "hv shape: %zu/%zu.\n", hv.rows(), hv.columns());
         if(nh_ != hv.columns()) {
             std::fprintf(stderr, "[%s] nh_: %u. hv.columns: %zu\n", __PRETTY_FUNCTION__, nh_, hv.columns());
@@ -397,14 +397,14 @@ public:
         const size_t nr = (~input).rows();
         const auto _l = l(), _k = k();
         for(unsigned i = 0; i < nr; ++i) {
-            auto r = row(hv, i, blz::unchecked);
+            auto r = row(hv, i, blaze::unchecked);
             for(unsigned j = 0; j < _l; ++j) {
                 insert(j, xxhasher_(&r[j * _k], sizeof(ElementType) * _k), idoffset + i);
             }
         }
     }
     template<typename VT, bool OSO>
-    shared::flat_hash_map<IT, unsigned> query(const blz::Vector<VT, OSO> &query) const {
+    shared::flat_hash_map<IT, unsigned> query(const blaze::Vector<VT, OSO> &query) const {
         auto hv = evaluate(hash(query));
         shared::flat_hash_map<IT, unsigned> ret;
         for(unsigned i = 0; i < l(); ++i) {
@@ -422,7 +422,7 @@ public:
     }
     template<typename MT, bool OSO>
     std::vector<shared::flat_hash_map<IT, unsigned>>
-    query(const blz::Matrix<MT, OSO> &query) const {
+    query(const blaze::Matrix<MT, OSO> &query) const {
         auto hv = evaluate(hash(query));
         //std::fprintf(stderr, "hv rows: %zu. columns: %zu. nh: %u. input num rows: %zu. input col: %zu\n", hv.rows(), hv.columns(), nh_, (~query).rows(), (~query).columns());
         if(hv.columns() != nh_) throw std::runtime_error("Wrong number of columns");
